+++ frontend/app/constants/message.ts
@@ -1,1 +1,0 @@
-export const SESSION_ID_ERROR = "Please fill out the form"